Communication costs by committee and candidates going - needs plain endpoint - needs by election - needs tests
