+++ webservices/resources/aggregates.py
@@ -306,57 +306,3 @@ class ScheduleBByPurposeView(BaseAggregateView):
     @schemas.marshal_with(schemas.ScheduleBByPurposePageSchema())
     def get(self, committee_id=None, **kwargs):
         return super().get(committee_id=committee_id, **kwargs)
-
-@spec.doc(
-    tags=['communicaiton_cost'],
-    description=(
-        'Communication cost by candidate aggregated by candidate ID.'
-    )
-)
-class CommunicationCostByCandidateView(BaseAggregateView):
-
-    model = models.CommunicationCostByCandidate
-    fields = [
-        ('cycle', models.CommunicationCostByCandidate.cycle),
-        ('candidate_id', models.CommunicationCostByCandidate.candidate_id),
-    ]
-
-    @args.register_kwargs(args.paging)
-    @args.register_kwargs(args.communicaion_cost_by_candidate)
-    @args.register_kwargs(
-        args.make_sort_args(
-            validator=args.IndexValidator(models.CommunicationCostByCandidate)
-        )
-    )
-    @schemas.marshal_with(schemas.CommunicationCostByCandidatePageSchema())
-    def get(self, committee_id=None, **kwargs):
-        return super().get(committee_id=committee_id, **kwargs)
-
-
-@spec.doc(
-    tags=['communicaiton_cost'],
-    path_params=[
-        utils.candidate_param,
-        utils.committee_param,
-    ],
-    description=(
-        'Communication cost by candidate aggregated by candidate ID and committee ID.'
-    ),
-)
-class CommunicationCostByCandidateView(BaseAggregateView):
-
-    model = models.CommunicationCostByCandidate
-    fields = [
-        ('cycle', models.CommunicationCostByCandidate.cycle),
-    ]
-
-    @args.register_kwargs(args.paging)
-    @args.register_kwargs(args.communicaion_cost_by_candidate)
-    @args.register_kwargs(
-        args.make_sort_args(
-            validator=args.IndexValidator(models.CommunicationCostByCandidate)
-        )
-    )
-    @schemas.marshal_with(schemas.CommunicationCostByCandidatePageSchema())
-    def get(self, committee_id=None, **kwargs):
-        return super().get(committee_id=committee_id, **kwargs)
